Refactor dataset configuration code
